+++ src/app/dashboard/results-analysis/results-analysis.component.ts
@@ -569,13 +569,13 @@ export class ResultsAnalysisComponent implements AfterViewInit {
   }
 
   public OnMappingDblClick(entry, event = null) {
-    if (entry instanceof AttackScenario) this.dialog.OpenAttackScenarioDialog(entry, false, this.dataSourceThreats.data);
+    if (entry instanceof AttackScenario) this.dialog.OpenAttackScenarioDialog(entry, false, this.dataSourceThreats.sortData(this.dataSourceThreats.filteredData, this.sortThreats));
     else if (entry instanceof Countermeasure) {
       if (event && event.target && this.displayedCountermeasureColumns[event.target.cellIndex] == 'progress' && entry.MitigationProcess) {
         this.dialog.OpenMitigationProcessDialog(entry.MitigationProcess, false);
       }
       else {
-        this.dialog.OpenCountermeasureDialog(entry, false, [], this.dataSourceCountermeasures.data);
+        this.dialog.OpenCountermeasureDialog(entry, false, [], this.dataSourceCountermeasures.sortData(this.dataSourceCountermeasures.filteredData, this.sortCountermeasures));
       }
     }
   }