+++ src/app/modeling/countermeasure-table/countermeasure-table.component.ts
@@ -166,7 +166,7 @@ export class CountermeasureTableComponent implements OnInit {
       let elements: ViewElementBase[] = null;
       if (this.selectedNode.data instanceof Diagram) elements = this.selectedNode.data.Elements.GetChildrenFlat();
       else if (this.selectedNode.data instanceof MyComponentStack) elements = this.selectedNode.data.GetChildrenFlat();
-      this.dialog.OpenCountermeasureDialog(entry, false, elements, [...this.dataSourceActive.data, ...this.dataSourceNA.data]);
+      this.dialog.OpenCountermeasureDialog(entry, false, elements, [...this.dataSourceActive.sortData(this.dataSourceActive.filteredData, this.sort), ...this.dataSourceNA.sortData(this.dataSourceNA.filteredData, this.sort)]);
     }
   }